+++ INWISE-SERVICES/src/main/java/com/prod/inwise/services/data/LineItem.java
@@ -1,5 +1,7 @@
 package com.prod.inwise.services.data;
 
+import java.math.BigDecimal;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.JoinColumn;
@@ -48,12 +50,12 @@ public class LineItem extends BaseModel {
 	@NotNull(message = "TotalTax can't be null")
 	@Column(name = "TOTAL_TAX", nullable = false)
 	@ApiModelProperty
-	private Float totalTax;
+	private BigDecimal totalTax;
 
 	@NotNull(message = "TotalPrice can't be null")
 	@Column(name = "TOTAL_PRICE", nullable = false)
 	@ApiModelProperty
-	private Float totalPrice;
+	private BigDecimal totalPrice;
 
 	/**
 	 * @return the invoice
@@ -114,28 +116,28 @@ public class LineItem extends BaseModel {
 	/**
 	 * @return the totalTax
 	 */
-	public Float getTotalTax() {
+	public BigDecimal getTotalTax() {
 		return totalTax;
 	}
 
 	/**
 	 * @param totalTax the totalTax to set
 	 */
-	public void setTotalTax(Float totalTax) {
+	public void setTotalTax(BigDecimal totalTax) {
 		this.totalTax = totalTax;
 	}
 
 	/**
 	 * @return the totalPrice
 	 */
-	public Float getTotalPrice() {
+	public BigDecimal getTotalPrice() {
 		return totalPrice;
 	}
 
 	/**
 	 * @param totalPrice the totalPrice to set
 	 */
-	public void setTotalPrice(Float totalPrice) {
+	public void setTotalPrice(BigDecimal totalPrice) {
 		this.totalPrice = totalPrice;
 	}
 }
